backend/requirements-dev.txt
django-stubs
ipdb
pytest
pytest-django
black